+++ src/App.vue
@@ -1,28 +1,40 @@
 <script setup>
-import {ref, watchEffect} from 'vue'
-import proverb from "@/assets/proverbs.json"
-const topic = ref('Amharic proverbs')
-const proverbs  = ref(proverb['Amharic proverbs'])
-function onSearch(item) {
-return item['English Equivalent Proverbs'].toLowerCase().includes('you')
+import {ref } from 'vue'
+
+let id = 0
+const newTodo = ref('')
+const todos = ref([
+ {id: id++, issue: 'Backlog'}, {id: id++, issue: 'Todo' },
+ {id: id++, issue: 'Going Hawaii'}
+])
+                              
+function addTodo() {
+ todos.value.push({id: id++,
+ issue: newTodo.value});
+ newTodo.value = '';
 }
-const filteredProverbs = ref([])
-
-watchEffect(() => {
-filteredProverbs.value = proverbs.value.filter(onSearch)
-})
-
+function removeTodo(item) {
+todos.value = todos.value.filter(todo => todo.id !== item.id) 
+}
 </script>
-
 <template>
-
-{{topic}}
-<ol> 
-<li v-for="(proverbEach, index) in filteredProverbs":key="index" >
-<pre>
-   {{proverbEach['Ge\'ez Writing']}} =  {{proverbEach['English Equivalent Proverbs']}}
-   
-</pre>
-</li>
-</ol>
+<form @submit.prevent="addTodo">
+ <input v-model="newTodo">
+  <button> add  </button>
+</form>
+<ul>
+ <li v-for = "todo in todos">
+  {{todo.issue}} - <button @click = "removeTodo(todo)"> remove  </button>
+ </li>
+</ul>
 </template>
+
+
+
+
+
+
+
+
+
+
